refactor: move reusable code to common module
